+++ app/src/main/java/com/inky/fitnesscalendar/db/Database.kt
@@ -29,7 +29,7 @@ import com.inky.fitnesscalendar.util.DATABASE_NAME
 private const val TAG = "Database"
 
 @Database(
-    version = 32,
+    version = 33,
     entities = [
         Activity::class,
         Recording::class,
@@ -61,6 +61,7 @@ private const val TAG = "Database"
         AutoMigration(from = 29, to = 30),
         AutoMigration(from = 30, to = 31),
         AutoMigration(from = 31, to = 32),
+        AutoMigration(from = 32, to = 33),
     ]
 )
 @TypeConverters(Converters::class)